# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for loading XKCD comics by number range and searching comics by query words.",
        "title": "XKCD Helper API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/last-comic": {
            "get": {
                "description": "Returns the last stored XKCD comic as a \"number, title, image URL\" object.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Get last comic",
                "responses": {
                    "200": {
                        "description": "Last comic",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchComic"
                        }
                    },
                    "500": {
                        "description": "Failed to send JSON",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/load-comics": {
            "get": {
                "description": "Returns XKCD comics as \"number, title\" strings for the requested range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Load comics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Starting comic number",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ending comic number",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loaded comics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LoadComic"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to send JSON",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Validates user credentials and returns a JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User login and password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON data or validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid login or password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/search-comics": {
            "get": {
                "description": "Renders a search form or XKCD comics matching the requested search query.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "comics"
                ],
                "summary": "Search comics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query words separated by spaces",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search page or matching comics page",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update": {
            "put": {
                "description": "Downloads and stores new XKCD comics that are not yet present in the database.",
                "tags": [
                    "comics"
                ],
                "summary": "Update comics",
                "responses": {
                    "200": {
                        "description": "Comics updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update comics",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.LoadComic": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "num": {
                    "type": "integer"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.SearchComic": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "num": {
                    "type": "integer"
                }
            }
        }
    }
}